Use GitHub compatible format for README
Renames README.txt to README.md, and updates format to use new parsable
format. README.md is moved to the 'create' template to simplify the
process of sample creation.

Also adds packaging.yaml for indexing, also in the 'create' template.

Finally, adds CONTRIB.md and LICENSE to 'base' template for compliance
purposes when the sample is distributed through GitHub.

Change-Id: Ib9b18fe4bf7a6f5b72f22b7d6137ba1964b9d3ea